third_party/WebKit/LayoutTests/svg/animations/animate-endElement-beginElement-expected.txt

CONSOLE WARNING: SVG's SMIL animations (<animate>, <set>, etc.) are deprecated and will be removed. Please use CSS animations or Web animations instead.
SVG 1.1 dynamic animation tests

Tests animation beginElement command's restarting capability after endElement.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rect.x.animVal.value is 50
PASS rect.x.baseVal.value is 0
PASS successfullyParsed is true

TEST COMPLETE
